+++ auto_route/example/lib/mobile/router/router.gr.dart
@@ -9,19 +9,18 @@
 // **************************************************************************
 
 import 'package:auto_route/auto_route.dart' as _i4;
-import 'package:example/mobile/screens/books/book_details_page.dart' as _i7;
-import 'package:example/mobile/screens/books/book_list_page.dart' as _i6;
-import 'package:example/mobile/screens/home_page.dart' as _i1;
-import 'package:example/mobile/screens/login_page.dart' as _i3;
-import 'package:example/mobile/screens/profile/my_books_page.dart' as _i9;
-import 'package:example/mobile/screens/profile/profile_page.dart' as _i8;
-import 'package:example/mobile/screens/settings.dart' as _i5;
-import 'package:example/mobile/screens/user-data/data_collector.dart' as _i2;
-import 'package:example/mobile/screens/user-data/single_field_page.dart'
-    as _i10;
-import 'package:example/mobile/screens/user-data/user_data_page.dart' as _i11;
-import 'package:flutter/foundation.dart' as _i13;
 import 'package:flutter/material.dart' as _i12;
+
+import '../screens/books/book_details_page.dart' as _i7;
+import '../screens/books/book_list_page.dart' as _i6;
+import '../screens/home_page.dart' as _i1;
+import '../screens/login_page.dart' as _i3;
+import '../screens/profile/my_books_page.dart' as _i9;
+import '../screens/profile/profile_page.dart' as _i8;
+import '../screens/settings.dart' as _i5;
+import '../screens/user-data/data_collector.dart' as _i2;
+import '../screens/user-data/single_field_page.dart' as _i10;
+import '../screens/user-data/user_data_page.dart' as _i11;
 
 class RootRouter extends _i4.RootStackRouter {
   RootRouter([_i12.GlobalKey<_i12.NavigatorState>? navigatorKey])
@@ -30,13 +29,13 @@ class RootRouter extends _i4.RootStackRouter {
   @override
   final Map<String, _i4.PageFactory> pagesMap = {
     HomeRoute.name: (routeData) {
-      return _i4.AdaptivePage<String>(
+      return _i4.MaterialPageX<String>(
           routeData: routeData, child: const _i1.HomePage());
     },
     UserDataCollectorRoute.name: (routeData) {
       final args = routeData.argsAs<UserDataCollectorRouteArgs>(
           orElse: () => const UserDataCollectorRouteArgs());
-      return _i4.AdaptivePage<_i2.UserData>(
+      return _i4.MaterialPageX<_i2.UserData>(
           routeData: routeData,
           child: _i2.UserDataCollectorPage(
               key: args.key, onResult: args.onResult));
@@ -44,7 +43,7 @@ class RootRouter extends _i4.RootStackRouter {
     LoginRoute.name: (routeData) {
       final args = routeData.argsAs<LoginRouteArgs>(
           orElse: () => const LoginRouteArgs());
-      return _i4.AdaptivePage<dynamic>(
+      return _i4.MaterialPageX<dynamic>(
           routeData: routeData,
           child: _i3.LoginPage(
               key: args.key,
@@ -52,34 +51,38 @@ class RootRouter extends _i4.RootStackRouter {
               showBackButton: args.showBackButton));
     },
     BooksTab.name: (routeData) {
-      return _i4.AdaptivePage<dynamic>(
+      return _i4.MaterialPageX<dynamic>(
           routeData: routeData, child: const _i4.EmptyRouterPage());
     },
     ProfileTab.name: (routeData) {
-      return _i4.AdaptivePage<dynamic>(
+      return _i4.MaterialPageX<dynamic>(
           routeData: routeData, child: const _i4.EmptyRouterPage());
     },
     SettingsTab.name: (routeData) {
-      final pathParams = routeData.pathParams;
+      final pathParams = routeData.inheritedPathParams;
+      final queryParams = routeData.queryParams;
       final args = routeData.argsAs<SettingsTabArgs>(
-          orElse: () => SettingsTabArgs(tab: pathParams.getString('tab')));
-      return _i4.AdaptivePage<dynamic>(
+          orElse: () => SettingsTabArgs(
+              tab: pathParams.getString('tab'),
+              query: queryParams.getString('query', 'none')));
+      return _i4.MaterialPageX<dynamic>(
           routeData: routeData,
-          child: _i5.SettingsPage(key: args.key, tab: args.tab));
+          child: _i5.SettingsPage(
+              key: args.key, tab: args.tab, query: args.query));
     },
     BookListRoute.name: (routeData) {
-      return _i4.AdaptivePage<dynamic>(
+      return _i4.MaterialPageX<dynamic>(
           routeData: routeData, child: _i6.BookListPage());
     },
     BookDetailsRoute.name: (routeData) {
-      final pathParams = routeData.pathParams;
+      final pathParams = routeData.inheritedPathParams;
       final args = routeData.argsAs<BookDetailsRouteArgs>(
           orElse: () => BookDetailsRouteArgs(id: pathParams.getInt('id', -1)));
-      return _i4.AdaptivePage<dynamic>(
+      return _i4.MaterialPageX<dynamic>(
           routeData: routeData, child: _i7.BookDetailsPage(id: args.id));
     },
     ProfileRoute.name: (routeData) {
-      return _i4.AdaptivePage<dynamic>(
+      return _i4.MaterialPageX<dynamic>(
           routeData: routeData, child: _i8.ProfilePage());
     },
     MyBooksRoute.name: (routeData) {
@@ -87,14 +90,14 @@ class RootRouter extends _i4.RootStackRouter {
       final args = routeData.argsAs<MyBooksRouteArgs>(
           orElse: () => MyBooksRouteArgs(
               filter: queryParams.optString('filter', 'none')));
-      return _i4.AdaptivePage<dynamic>(
+      return _i4.MaterialPageX<dynamic>(
           routeData: routeData,
           child: _i9.MyBooksPage(key: args.key, filter: args.filter));
     },
     NameFieldRoute.name: (routeData) {
       final args = routeData.argsAs<NameFieldRouteArgs>(
           orElse: () => const NameFieldRouteArgs());
-      return _i4.AdaptivePage<dynamic>(
+      return _i4.MaterialPageX<dynamic>(
           routeData: routeData,
           child: _i10.SingleFieldPage(
               key: args.key,
@@ -105,7 +108,7 @@ class RootRouter extends _i4.RootStackRouter {
     FavoriteBookFieldRoute.name: (routeData) {
       final args = routeData.argsAs<FavoriteBookFieldRouteArgs>(
           orElse: () => const FavoriteBookFieldRouteArgs());
-      return _i4.AdaptivePage<dynamic>(
+      return _i4.MaterialPageX<dynamic>(
           routeData: routeData,
           child: _i10.SingleFieldPage(
               key: args.key,
@@ -116,7 +119,7 @@ class RootRouter extends _i4.RootStackRouter {
     UserDataRoute.name: (routeData) {
       final args = routeData.argsAs<UserDataRouteArgs>(
           orElse: () => const UserDataRouteArgs());
-      return _i4.AdaptivePage<dynamic>(
+      return _i4.MaterialPageX<dynamic>(
           routeData: routeData,
           child: _i11.UserDataPage(key: args.key, onResult: args.onResult));
     }
@@ -181,7 +184,7 @@ class HomeRoute extends _i4.PageRouteInfo<void> {
 class UserDataCollectorRoute
     extends _i4.PageRouteInfo<UserDataCollectorRouteArgs> {
   UserDataCollectorRoute(
-      {_i13.Key? key,
+      {_i12.Key? key,
       dynamic Function(_i2.UserData)? onResult,
       List<_i4.PageRouteInfo>? children})
       : super(name,
@@ -195,15 +198,20 @@ class UserDataCollectorRoute
 class UserDataCollectorRouteArgs {
   const UserDataCollectorRouteArgs({this.key, this.onResult});
 
-  final _i13.Key? key;
+  final _i12.Key? key;
 
   final dynamic Function(_i2.UserData)? onResult;
+
+  @override
+  String toString() {
+    return 'UserDataCollectorRouteArgs{key: $key, onResult: $onResult}';
+  }
 }
 
 /// generated route for [_i3.LoginPage]
 class LoginRoute extends _i4.PageRouteInfo<LoginRouteArgs> {
   LoginRoute(
-      {_i13.Key? key,
+      {_i12.Key? key,
       void Function(bool)? onLoginResult,
       bool showBackButton = true})
       : super(name,
@@ -220,11 +228,16 @@ class LoginRouteArgs {
   const LoginRouteArgs(
       {this.key, this.onLoginResult, this.showBackButton = true});
 
-  final _i13.Key? key;
+  final _i12.Key? key;
 
   final void Function(bool)? onLoginResult;
 
   final bool showBackButton;
+
+  @override
+  String toString() {
+    return 'LoginRouteArgs{key: $key, onLoginResult: $onLoginResult, showBackButton: $showBackButton}';
+  }
 }
 
 /// generated route for [_i4.EmptyRouterPage]
@@ -245,21 +258,29 @@ class ProfileTab extends _i4.PageRouteInfo<void> {
 
 /// generated route for [_i5.SettingsPage]
 class SettingsTab extends _i4.PageRouteInfo<SettingsTabArgs> {
-  SettingsTab({_i13.Key? key, required String tab})
+  SettingsTab({_i12.Key? key, required String tab, String query = 'none'})
       : super(name,
             path: 'settings/:tab',
-            args: SettingsTabArgs(key: key, tab: tab),
-            rawPathParams: {'tab': tab});
+            args: SettingsTabArgs(key: key, tab: tab, query: query),
+            rawPathParams: {'tab': tab},
+            rawQueryParams: {'query': query});
 
   static const String name = 'SettingsTab';
 }
 
 class SettingsTabArgs {
-  const SettingsTabArgs({this.key, required this.tab});
+  const SettingsTabArgs({this.key, required this.tab, this.query = 'none'});
 
-  final _i13.Key? key;
+  final _i12.Key? key;
 
   final String tab;
+
+  final String query;
+
+  @override
+  String toString() {
+    return 'SettingsTabArgs{key: $key, tab: $tab, query: $query}';
+  }
 }
 
 /// generated route for [_i6.BookListPage]
@@ -284,6 +305,11 @@ class BookDetailsRouteArgs {
   const BookDetailsRouteArgs({this.id = -1});
 
   final int id;
+
+  @override
+  String toString() {
+    return 'BookDetailsRouteArgs{id: $id}';
+  }
 }
 
 /// generated route for [_i8.ProfilePage]
@@ -295,7 +321,7 @@ class ProfileRoute extends _i4.PageRouteInfo<void> {
 
 /// generated route for [_i9.MyBooksPage]
 class MyBooksRoute extends _i4.PageRouteInfo<MyBooksRouteArgs> {
-  MyBooksRoute({_i13.Key? key, String? filter = 'none'})
+  MyBooksRoute({_i12.Key? key, String? filter = 'none'})
       : super(name,
             path: 'my-books',
             args: MyBooksRouteArgs(key: key, filter: filter),
@@ -307,15 +333,20 @@ class MyBooksRoute extends _i4.PageRouteInfo<MyBooksRouteArgs> {
 class MyBooksRouteArgs {
   const MyBooksRouteArgs({this.key, this.filter = 'none'});
 
-  final _i13.Key? key;
+  final _i12.Key? key;
 
   final String? filter;
+
+  @override
+  String toString() {
+    return 'MyBooksRouteArgs{key: $key, filter: $filter}';
+  }
 }
 
 /// generated route for [_i10.SingleFieldPage]
 class NameFieldRoute extends _i4.PageRouteInfo<NameFieldRouteArgs> {
   NameFieldRoute(
-      {_i13.Key? key,
+      {_i12.Key? key,
       String message = '',
       String willPopMessage = '',
       void Function(String)? onNext})
@@ -334,20 +365,25 @@ class NameFieldRouteArgs {
   const NameFieldRouteArgs(
       {this.key, this.message = '', this.willPopMessage = '', this.onNext});
 
-  final _i13.Key? key;
+  final _i12.Key? key;
 
   final String message;
 
   final String willPopMessage;
 
   final void Function(String)? onNext;
+
+  @override
+  String toString() {
+    return 'NameFieldRouteArgs{key: $key, message: $message, willPopMessage: $willPopMessage, onNext: $onNext}';
+  }
 }
 
 /// generated route for [_i10.SingleFieldPage]
 class FavoriteBookFieldRoute
     extends _i4.PageRouteInfo<FavoriteBookFieldRouteArgs> {
   FavoriteBookFieldRoute(
-      {_i13.Key? key,
+      {_i12.Key? key,
       String message = '',
       String willPopMessage = '',
       void Function(String)? onNext})
@@ -366,18 +402,23 @@ class FavoriteBookFieldRouteArgs {
   const FavoriteBookFieldRouteArgs(
       {this.key, this.message = '', this.willPopMessage = '', this.onNext});
 
-  final _i13.Key? key;
+  final _i12.Key? key;
 
   final String message;
 
   final String willPopMessage;
 
   final void Function(String)? onNext;
+
+  @override
+  String toString() {
+    return 'FavoriteBookFieldRouteArgs{key: $key, message: $message, willPopMessage: $willPopMessage, onNext: $onNext}';
+  }
 }
 
 /// generated route for [_i11.UserDataPage]
 class UserDataRoute extends _i4.PageRouteInfo<UserDataRouteArgs> {
-  UserDataRoute({_i13.Key? key, dynamic Function(_i2.UserData)? onResult})
+  UserDataRoute({_i12.Key? key, dynamic Function(_i2.UserData)? onResult})
       : super(name,
             path: 'results',
             args: UserDataRouteArgs(key: key, onResult: onResult));
@@ -388,7 +429,12 @@ class UserDataRoute extends _i4.PageRouteInfo<UserDataRouteArgs> {
 class UserDataRouteArgs {
   const UserDataRouteArgs({this.key, this.onResult});
 
-  final _i13.Key? key;
+  final _i12.Key? key;
 
   final dynamic Function(_i2.UserData)? onResult;
+
+  @override
+  String toString() {
+    return 'UserDataRouteArgs{key: $key, onResult: $onResult}';
+  }
 }