+++ auto_route/lib/src/route/route_data.dart
@@ -75,12 +75,10 @@ class RouteData {
   String get match => _match.stringMatch;
 
   Parameters get inheritedPathParams {
-    if (_parent == null) {
-      return const Parameters(const {});
-    }
-    return _parent!.breadcrumbs.map((e) => e.pathParams).reduce(
+    final params = breadcrumbs.map((e) => e.pathParams).reduce(
           (value, element) => value + element,
         );
+    return params;
   }
 
   Parameters get pathParams => _match.pathParams;
@@ -110,8 +108,21 @@ class RouteData {
       identical(this, other) ||
       other is RouteData &&
           runtimeType == other.runtimeType &&
-          route == other.route;
+          _match == other._match &&
+          _parent == other._parent;
 
   @override
-  int get hashCode => route.hashCode;
+  int get hashCode => _match.hashCode ^ _parent.hashCode;
+
+// @override
+  // bool operator ==(Object other) =>
+  //     identical(this, other) ||
+  //     other is RouteData &&
+  //         runtimeType == other.runtimeType &&
+  //         route == other.route;
+  //
+  //
+  //
+  // @override
+  // int get hashCode => route.hashCode;
 }