+++ auto_route/lib/src/router/controller/auto_route_guard.dart
@@ -20,11 +20,13 @@ abstract class AutoRouteGuard {
 }
 
 class NavigationResolver {
+  final StackRouter _router;
   final Completer<bool> _completer;
   final RouteMatch route;
   final List<RouteMatch> pendingRoutes;
 
   NavigationResolver(
+    this._router,
     this._completer,
     this.route, {
     this.pendingRoutes = const [],
@@ -38,7 +40,8 @@ class NavigationResolver {
   bool get isResolved => _completer.isCompleted;
 }
 
-abstract class AutoRedirectGuard extends AutoRouteGuard with ChangeNotifier {
+abstract class AutoRedirectGuardBase extends AutoRouteGuard
+    with ChangeNotifier {
   late ReevaluationStrategy _strategy;
 
   ReevaluationStrategy get strategy => _strategy;
@@ -56,15 +59,44 @@ abstract class AutoRedirectGuard extends AutoRouteGuard with ChangeNotifier {
     notifyListeners();
   }
 
-  void _reevaluate(StackRouter stackRouter) {
+  Future<void> _reevaluate(StackRouter stackRouter) async {
     _strategy.reevaluate(this, stackRouter);
+  }
+}
+
+abstract class AutoRedirectGuard extends AutoRedirectGuardBase {
+  NavigationResolver? _redirectResolver;
+  @protected
+  void redirect(PageRouteInfo route,
+      {required NavigationResolver resolver}) async {
+    if (_redirectResolver != null) return;
+
+    _redirectResolver = resolver;
+    assert(!resolver.isResolved, 'Resolver is already completed');
+    final router = resolver._router._findStackScope(route);
+    router.push(route);
+    await resolver._completer.future;
+    if (router.current.name == route.routeName) {
+      router.removeLast();
+      router.markUrlStateForReplace();
+    }
+    _redirectResolver = null;
+  }
+
+  @override
+  Future<void> _reevaluate(StackRouter stackRouter) async {
+    if (_redirectResolver != null) {
+      onNavigation(_redirectResolver!, stackRouter);
+    } else {
+      super._reevaluate(stackRouter);
+    }
   }
 }
 
 abstract class ReevaluationStrategy {
   const ReevaluationStrategy._();
 
-  void reevaluate(AutoRedirectGuard guard, StackRouter router);
+  void reevaluate(AutoRedirectGuardBase guard, StackRouter router);
 
   const factory ReevaluationStrategy.rePushAllRoutes() = RePushAllStrategy;
 
@@ -82,7 +114,7 @@ class RePushAllStrategy extends ReevaluationStrategy {
   const RePushAllStrategy() : super._();
 
   @override
-  void reevaluate(AutoRedirectGuard guard, StackRouter router) {
+  void reevaluate(AutoRedirectGuardBase guard, StackRouter router) {
     final stackData = router.stackData;
     final routesToRemove =
         List<RouteMatch>.unmodifiable(stackData.map((e) => e._match));
@@ -98,7 +130,7 @@ class RePushFirstGuarded extends ReevaluationStrategy {
   const RePushFirstGuarded() : super._();
 
   @override
-  void reevaluate(AutoRedirectGuard guard, StackRouter router) {
+  void reevaluate(AutoRedirectGuardBase guard, StackRouter router) {
     final routes = router.stackData.map((e) => e.route).toList();
     final firstGuardedRouteIndex =
         routes.indexWhere((r) => r.guards.contains(guard));
@@ -117,7 +149,7 @@ class RePushFirstGuardedAndUp extends ReevaluationStrategy {
   const RePushFirstGuardedAndUp() : super._();
 
   @override
-  void reevaluate(AutoRedirectGuard guard, StackRouter router) {
+  void reevaluate(AutoRedirectGuardBase guard, StackRouter router) {
     final routes = router.stackData.map((e) => e.route).toList();
     final firstGuardedRouteIndex =
         routes.indexWhere((r) => r.guards.contains(guard));
@@ -138,7 +170,7 @@ class _removeFirstGuardedAndUp extends ReevaluationStrategy {
   const _removeFirstGuardedAndUp() : super._();
 
   @override
-  void reevaluate(AutoRedirectGuard guard, StackRouter router) {
+  void reevaluate(AutoRedirectGuardBase guard, StackRouter router) {
     final routes = router.stackData.map((e) => e.route).toList();
     final firstGuardedRouteIndex =
         routes.indexWhere((r) => r.guards.contains(guard));