+++ auto_route_generator/CHANGELOG.md
@@ -1,4 +1,14 @@
 # ChangeLog
+## [3.1.0]
+- Prefer previous-current route index as new current index when updating tab routes #797
+- Improve native navigation history
+- Use browser history as navigation history in web
+- Fix path/query params not updating issue #809
+- Fix generic nullable types are generated as non-nullables #811
+- Override toString method inside of generated arg classes [FR] #820
+- Support redirect paths with path params [FR] #818
+- Fix replace/replaceNamed/replaceAll do not replace current url in browser #781
+
 ## [3.0.1]
 - Fix conflict with source_gen:combining_builder by using .gr.dart instead of .g.dart
 ## [3.0.0]